Split AccountController into Public, Authenticated and Admin for better project structure
