+++ api/src/main/java/SpringProject/_Spring/controller/AccountController/AccountControllerAdmin.java
@@ -1,0 +1,21 @@
+package SpringProject._Spring.controller.AccountController;
+
+import SpringProject._Spring.service.AccountService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api")
+public class AccountControllerAdmin {
+
+    private final AccountService accountService;
+    private final PasswordEncoder passwordEncoder;
+
+    @Autowired
+    public AccountControllerAdmin(AccountService accountService, PasswordEncoder passwordEncoder) {
+        this.accountService = accountService;
+        this.passwordEncoder = passwordEncoder;
+    }
+}
