+++ api/src/test/java/SpringProject/_Spring/AccountControllerTest/AccountJunitPUTTest.java
@@ -1,6 +1,6 @@
 package SpringProject._Spring.AccountControllerTest;
 
-import SpringProject._Spring.controller.AccountController;
+import SpringProject._Spring.controller.AccountController.AccountControllerAuthenticated;
 import SpringProject._Spring.dto.PasswordUpdateDTO;
 import SpringProject._Spring.model.Account;
 import SpringProject._Spring.model.Role;
@@ -23,14 +23,13 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.when;
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import java.util.List;
 import java.util.Optional;
 
-@WebMvcTest(controllers = AccountController.class)
+@WebMvcTest(controllers = AccountControllerAuthenticated.class)
 @Import(SecurityConfig.class)
 @AutoConfigureMockMvc
 public class AccountJunitPUTTest {
